Better code organization and separation.
