+++ timber/helpers.py
@@ -1,6 +1,5 @@
 # coding: utf-8
 from __future__ import print_function, unicode_literals
-import json
 
 
 class TimberContext(object):
@@ -32,11 +31,3 @@ class TimberContext(object):
             for name, data in contexts.items():
                 x.setdefault(name, {}).update(data)
         return x
-
-
-def make_context(*args, **kwargs):
-    return TimberContext(*args, **kwargs)
-
-
-def _debug(*data):
-    print(json.dumps(data, indent=2, sort_keys=True))
